# openapi/media-svc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/playlist-items": {
      "post": {
        "operationId": "PlaylistItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistItem"
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlaylistItemController_findAll",
        "parameters": [
          {
            "name": "text",
            "required": false,
            "in": "query",
            "allowEmptyValue": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "explode": true,
            "allowEmptyValue": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/playlist-items/{playlistItemId}": {
      "put": {
        "operationId": "PlaylistItemController_update",
        "parameters": [
          {
            "name": "playlistItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistItem"
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PlaylistItemController_remove",
        "parameters": [
          {
            "name": "playlistItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PlaylistItemController_findOne",
        "parameters": [
          {
            "name": "playlistItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistItemDto"
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/playlist-items/{playlistItemId}/share/{userId}": {
      "post": {
        "operationId": "PlaylistItemController_share",
        "parameters": [
          {
            "name": "playlistItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": ":userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareItem"
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/playlist-items/popular": {
      "get": {
        "operationId": "PlaylistItemController_findPopular",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["playlist-items"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/playlists/{playlistId}/share/{userId}": {
      "post": {
        "operationId": "PlaylistController_share",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareItem"
                  }
                }
              }
            }
          }
        },
        "tags": ["playlists"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Media Service",
    "description": "Media Service",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local"
    },
    {
      "url": "https://mediashare-api-staging.herokuapp.com",
      "description": "staging"
    },
    {
      "url": "https://mediashare-api-prod.herokuapp.com",
      "description": "production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TagKeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["key", "value"]
      },
...
      "MediaVisibilityType": {
        "type": "string",
        "enum": ["private", "shared", "subscription", "public"]
      },
...
      "ShareItem": {
        "type": "object",
        "properties": {}
      },
...
      "PlaylistItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "playlistId": {
            "type": "string",
            "example": "ulCPNGrw8gdyQpoRlVpkPuIVm9cXkrdIOcLwYFBdgoMyWlO0rCopf3kw2WYPRZYSWWBRFW6u02i2jAVDm99O8hhZL0SuyUk9JJ7Hs8JJiCWKZJ04oY2SepFCcKsUv",
            "maxLength": 255,
            "minLength": 0
          },
          "mediaId": {
            "type": "string",
            "example": "KOh0ONnoPho1J9av5qJ7zQ9KZkRmiKsUU7B8AlSSUKEh1zMQn8u8vVe2VZe8E3tNXNKaYfewdQ4LdNAukhfKBbDAnLsMDgAmZuVtGhSzCSOAL0n0GzqDuWyY8VL6m",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "CT4z7jgucsC4CKSyB2NmjfJPlIcoMhnUt9nGIP0HUq1CD1sNS0v0vaDPX5Z36t5h6WsrxpeYV8Dh1IfzWvnmmpZgMGL2UU3ZbS96ZOQePVeOzFTHcuV09yl5p0J0h",
            "maxLength": 255,
            "minLength": 0
          },
          "createdBy": {
            "type": "string",
            "example": "cHyTIgZ3Do36F024TX20K6BgPsHKGxSEa74ulZyqEd8xf33dgCBZEXdBp1fRVMzybMZiuKkUUcbt4AmAiLTj9rgjrvWap8KmaPnIWTh7GyTzuR9Lf3W1FmaqLlPym",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "iPNxxMq01qJ4wZYZpVQraWzkGlhjbVAtdvjFtV9gEwIzYGktaPywSUFWnFUBXs3ZkxlE9GzKzYyjgjkcIqo0XeoSCeih7g2JANvzXse4TXhmScZtfRPP1CiUJ3rpF",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "SjS4DKvbyU4VJi2jW24W9CYTkoD1d4zLSi3kED7jKmrfnAR5EgkIrELmj4XY0o9oi22Uaq3HmTVkaIZULvbnZnvQIDQ1iUUjXtEdKt6dIZPcdZnF2wCKF971AKy7ePV1TtlmE9MXfftBRIf08DLrYsI0FjMA3ChcE0snBGefguujONGAstjTUnDWG1kZN4Itf9RGmniisuzZaoo9Ev4zTnIwrJllhyNErwXHL1gdDwgjpwWiy9QiJfSh2Ib6j5W7VWBNOM1pnhioLUuwtfiSE5ElLpM8J8H8P8pD29ppEZmW"
          },
          "description": {
            "type": "string",
            "example": "fovQvqJGC9YreB7m3Kk9SMw29J1HulFANsvb1L1Qblt9BrXrPhCpW1AFCK3HZ3CVRqWDwKApKVbbiZl440SR6NAEu3D5hU2n1mL02yYbSpRXayYfNNI6cXXLX2LvBrjarbr2S2f9yi58eTva9n3cKK2GBOrFDu68itIPmfbj9QtJNa08EgglNOnvf8bPttW2B1g2Ohjxmj8E4IvpZUYWiNB5GENzAhQ8cuoFfdK73ReWK0TrpmNx3U4IrqmerqtLpFrERfTxm7SQXqIByy3fvOIWUgIg82HOpUttzuik12kD"
          },
          "uri": {
            "type": "string",
            "example": "f0nVvMsjNPXiBQ8eONYheyc9ZTPpdhJKVCzG3lPJHfqrXZYBqrIpdxBpC0i4ERkLugtvWcbKjunkLS2mPGdm5s2mocCwS3A2d5vaj6xK1tTUU6yR3CUJMIFVbRuIp91Z75tlF43PHr61q0ANurcz9qU7ohGK1M3DzU8NcHyoavQ2yrdVyUdFaOluFu7HcQpmBzRktgBMJtbPqRpgeGoUnyN9imcPGM2KvLFhF1npxVhthFp51svj6YnO69MEfiy4pa37a6sMaF1405JHbuopfG9VqswpbuQsPexduX7wcaOD"
          },
          "imageSrc": {
            "type": "string",
            "example": "pTmJhZaoJLq2HdQmyKW2TxEvGOUJUTC0iSZ6Fl0yKoBbaohyldvS0VWvPsBSLFM5iXBKBmUqqbz6nZAI47mim4JNnkvT6YEWFUEd3KBGNpRrX371nDs0V7AHoU4G3JgszTYgyabued4pc3mYcpkWWkGNYctikMshWrvssHdWKRqIWI8WDvZStxqAPlP6Gv5nNTQ3NFe5S5NBfoH9PalEor7NVnOupXv9xU2S7i88dtKFfRHMieCQK7HWKdJprCFjb0spvmaVubCA47vnqYquvRAk2Le7Hn7MaNLnkEkECuLg"
          },
          "sortIndex": {
            "type": "number"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          },
          "shareCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "playlistId",
          "mediaId",
          "userId",
          "createdBy",
          "title",
          "description",
          "uri",
          "visibility",
          "shareCount",
          "viewCount",
          "likesCount"
        ]
      },
...
      "CreatePlaylistItemDto": {
        "type": "object",
        "properties": {
          "playlistId": {
            "type": "string",
            "example": "enF8Ju1HQvZxlhtac3xoFLry678seB6iN4XO8sF0wjTk5pQ10meBrixMVXMXhaqkawBQLt83YBlpY7i3ylzQCOly0w2fJu3Ffx4gdMZ5mVVJI8SJkkg8sJ5DUEJQH",
            "maxLength": 255,
            "minLength": 0
          },
          "mediaId": {
            "type": "string",
            "example": "AgCHNhhTbCO8jdmA68FbM9SUxbBGXy6FmdWyEc74h7rv1eLupU8EkK5KvFoxVHQvwSjAWkhbG9UfYwKW5TuOu9F9jBSlXXpRes2UuWmElQ0X3a4omg1SATvD57BzW",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "I5bjD0yN0NGjH2u1Ug4Gyr9sSZbvgAhUhvwJhRUxX4TJvH7DlAwYlLL6BETrLJaYHyPE1kaZVA4XtKDj2di6lNRGaQQRkiBTdBcngko5fLcTSFQ8YPQTbkT9IWeiv",
            "maxLength": 255,
            "minLength": 0
          },
          "createdBy": {
            "type": "string",
            "example": "aLzDWWtOzjL1YPesEF3OPf3uOghOOj59UHzppavRlXP283WBCmtyuqoMLq3eAyyD4siFjFflFJG8C4bihXYmpiwkfcJUegvbd6sl2jLDMuTKV85JCA0Jd7JkExZK4",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "YdVMYGbGTxKvYYpkYHjTMqyDpeeu1LM0caidmAHgoktTsIC1pbyL9P6Q9KWrxf2xKZ4MFN48Z7JwAvEGmrUCNoNX5ayv0SvmYfXFNLLEcSr8aWqfP5yUi494FJMiQ",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "75EGc8JlG297cJ3oRV84DI5cEGNNM6h3RQJk1D932hH1aAqMQuXswWGVFfX8i4lbkN0f0xAA3IriXKzdDYMDcEwsAFA7zB6vGR9glnGwvZEOLQ9FJoFuI0IdMsaCAUE9guYpZ5RSaZUEfAtYKDPgPKUrnYqa1z8A8ipJzgFH4fJA3N1KX4kebLqn48Mmf72i08XZXMQAqwKzQ61nfXjqL7FmOXXR3dlcFg7Ze7Z3E6l2l0XC8uOz3Zvh5jtUW7z9iIuatg7ODuraoOMCkrMBk4LXQ4xQqdYK2qqREBMd4lsE"
          },
          "description": {
            "type": "string",
            "example": "ONQuukmJXFrXSiV7Oa6xfOkSBokf2VG9DOGXU4LwzREiB1vkLZt1kDcAWdOKigAc3dhp5BPgOB7BrW14ruUUOayz30w6vPPLaHrZ1ZTKwbwjN3a6Pl14xAbbb0KfQw0zSsamyvpaAqLiM1PxyIiv072JGfLImYltCZlHqrjTU1SanxzAO3PT8yIBDP1Vr2kR0d3z2CdVIDmeRHE9EiexxUolNBatvHUsCY9X4IwHWAUgmXX3WBPGCDYjAKCKWBhLzP7QxxCrwIZQ2kJDBoBxPDAAENqedDdcfUZ7luOdFehP"
          },
          "uri": {
            "type": "string",
            "example": "OnLtb86e9mLEgUiVT8X1oncLKVnFv0O9DexgRijxiPDwmitooQnLT3OfUckxW3pPbbZBt2gB8D4nDAR4du41XXfz8YIJ31N8JwgREXAEeiYUa2Hab7QBCKPpwc41LeOTLnZTDvQyCHwF38wYrJ1rrLmstdhfL2sYghhazfpCkuPS9FsdqSi6IJ9RUyzK29kWrS6ibUC5j2fMn64jxxpkxQh5keCp1kFpTBY3EIZTWIzBeopoaoMaobvug5DApeXAkOgL9J3w7q0cuRJzhoaWEosK9nBxSHBJo8HwrQ2xsVdb"
          },
          "imageSrc": {
            "type": "string",
            "example": "zI437AhlS5K6JE0kVX73lZQVRDMaQUCNhsw11mVAO6EG8TzYji99EUsJRCPV0WiMNT1ldczZWBLmfXNgBjHPCzJGjEzoiAFxHLLcu6sOsxQCl6yOhBquXEsmXrx2WIsynEkaLzjXUtDyggWwOfkC5M2PeDVpUnQTTcYFAGibwolGyo00pqoKbDeKmT4CLqSJRo8wA294Kq2bu0vsd5leBKlhEuIsx1SMVSerUROBUXOtMKsBHPcWLWPFGjpRhDnBk10KpTu9Ra4rf75HOnD3rnenLCe0JMRIaSIkBdMB5BZ6"
          },
          "sortIndex": {
            "type": "number"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          }
        },
        "required": [
          "playlistId",
          "mediaId",
          "userId",
          "createdBy",
          "title",
          "description",
          "uri",
          "visibility"
        ]
      },
      "PlaylistItem": {
        "type": "object",
        "properties": {}
      },
      "UpdatePlaylistItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "JWuZMCJwXFLK53lVTQ8CJiC7IWxujnylov4h9AwYvu2ar8O2a0TlJiERkTMKGcx3q4hsBsaoqVHMmrFLs25NmF7YGG51ZOZ9VTTeeMJsJkPxVvFH03CMohFcaz2bp",
            "maxLength": 255,
            "minLength": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "playlistId": {
            "type": "string",
            "example": "tLvL0DFENlFW9m5bj7dffPhbhO72rSvalAdF0WcfyOikD7ZULLCb5SuemtJsjCIoKCLOdmS18W1yxiLzYh845yYyPolNcmJzQTa80IMOdASrpzkK6AQkx9zdJ8KKQ",
            "maxLength": 255,
            "minLength": 0
          },
          "mediaId": {
            "type": "string",
            "example": "711SyMf3fDvMKDWOxJixTG8IfB20tCroJaZHnxdgXgQqb77DIIfgZHrsHn8auWPar1X2PuuDc8tpftSiOgGbwQvDQYpAuPnoU1JU8oSAlwNF2etWwzcfZXxTjFJA7",
            "maxLength": 255,
            "minLength": 0
          },
          "userId": {
            "type": "string",
            "example": "leXIxQ4z0mrhvCcoQeIA4OUGV9KBtVU9QXEpJpAqq6l7mNriFE8GfSrSEEVILYqA7P3b4S0Wb44yojr9qEsQAsSbEM0MwtCizaZ1qCGIVsCJewkzcnIH3kkAms3yQ",
            "maxLength": 255,
            "minLength": 0
          },
          "createdBy": {
            "type": "string",
            "example": "XBDR9QdO3BDqNdv29bIemqERo0kO5BkHmDB2s27Jv3lHakdQatRR1XD1Kev7gQDdJbTnUQnCOUzMFbKM3zUh462VfnnNVuNmNaTmagzvfc7YYBEouyjx2TzV81Nng",
            "maxLength": 255,
            "minLength": 0
          },
          "title": {
            "type": "string",
            "example": "k37nNatvlyvKpIHEo8xETPUZYIhW8B4h6vB54Zfu8luaoDBgJrtHWtl94maT9kVgmjTrfxpefL9EPlwgysk29In1DgXBkGoe7vW18piFUEHCv7ELHPyZ9L8BZtfeh",
            "maxLength": 255,
            "minLength": 0
          },
          "summary": {
            "type": "string",
            "example": "IY5v1dDVt0XplmTiYbV1jyIIrWAdjPzkRGakt8HayAVSJATPExeDaiGolnxP8IjMMWRkjx4a675bBKr9d2Hu6Nyd1Kw94UMVuMtUyDkqXRla91td7YR8MmISMydjyGWL3OuHviBKciM0IjtsgcyLFrxwgRBE7yOGglFUSTEY3fuxX7M9SZHYmDHDoxuIFMcsdUJBKiGm7NX83PAnQijV8I0NiUMN8eJMnUwZrBchsEg2NnCDBKG4kdLc51qsu6E2vc2x6u4wrcfIaDQMFUSeoILZpfygAGQDf4SRmh58JvTY"
          },
          "description": {
            "type": "string",
            "example": "LDczGNCTJrMYPEjy6j76JwddcoafTW9ZwlP2HZc1ZT3IPSm4pCqhUgsrbClRHCqRgMQa29gXdDR1KuSJ6wS1i5qCkcLKqmE5Q3TLm84P05MvcRI6wcgvcO7fDrJAaNtV70qBRLOULQiVAP3U5xNYrGtsbJU5EWEvKXTGpm82wTnCDTgucKXJpfYd7Riiy1rOQT1XkF2vak7IdiFxudM9BNyO2heHmrngto5w2RLcAHn5ZXjHzxqXPy1Wq8fUMK17SzA6rIpPVK3IMbBVFZxkt9qOQbdJlngOXO7Vd3HRF9iN"
          },
          "uri": {
            "type": "string",
            "example": "3knDQXMGpoB2kncWdkQoZZE1PS93wFx6uZNivwRtgxbS0onDsC62rEnn9yqHDfqdH7XV294bUhaF0HiEKrfWifF28sYqDZlnsUxBUqnVKB4I8cLOZiY9gBlCcnvotZJxOQCg01n2guHqFW9jNz7tILcof2dojtJwTqzLkYg60jeyiLlxCqd1aEgJcoz1oogxqQEn4uRAj5TiOtY2c2gbdZ29OU6wX6lDnJNOaEHcJYYlkSIkDNyd8eIXQinnTxGPyYXeY2tucn2pme4KhilYUJUu0Wg1sjsufx2qYOkpPMyn"
          },
          "imageSrc": {
            "type": "string",
            "example": "oRqk8NNogCPCYKEnvRaCOiYkNtCEVnC2Jy25kmxE4ZRoC2IYFe52m68wfmolvrKzYqRwG8nOy1pDNHkRu6qJfX8RTCpNOxXHawiz9G5e1nkbhgkh5O04jvJ7QD25CHyeil2GTFUTYrgUuxUDtswWp9OAm1KGMqdcyzyZzJTRicXF6gpnVNGlrUoItevKx2DostSeCFxnw8E2ED7hEnLbaI5DPFRklWe9XgdmnJzjWBXD0LKbosaSCZ8oVzC6EAFi7MFywcE0ZYGoZgn3iFnUZAlahWx8fYrcn35trJjK2KEu"
          },
          "sortIndex": {
            "type": "number"
          },
          "isPlayable": {
            "type": "boolean"
          },
          "visibility": {
            "$ref": "#/components/schemas/MediaVisibilityType"
          },
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagKeyValue"
            }
          }
        },
        "required": [
          "_id",
          "playlistId",
          "mediaId",
          "userId",
          "createdBy",
          "title",
          "description",
          "uri"
        ]
      },
...
    }
  }
}